+++ x68k/irqh.c
@@ -15,7 +15,7 @@ typedef int32_t  FASTCALL C68K_INT_CALLBACK(int32_t level);
 typedef int32_t  FASTCALL C68K_INT_CALLBACK(int32_t level);
 #endif
 
-	BYTE	IRQH_IRQ[8];
+	uint8_t	IRQH_IRQ[8];
 	void	*IRQH_CallBack[8];
 
 // -----------------------------------------------------------------------
@@ -30,10 +30,10 @@ void IRQH_Init(void)
 // -----------------------------------------------------------------------
 //   デフォルトのベクタを返す（これが起こったら変だお）
 // -----------------------------------------------------------------------
-DWORD FASTCALL IRQH_DefaultVector(BYTE irq)
+uint32_t FASTCALL IRQH_DefaultVector(uint8_t irq)
 {
 	IRQH_IRQCallBack(irq);
-	return -1;
+	return (uint32_t)-1;
 }
 
 
@@ -41,7 +41,7 @@ DWORD FASTCALL IRQH_DefaultVector(BYTE irq)
 //   他の割り込みのチェック
 //   各デバイスのベクタを返すルーチンから呼ばれます
 // -----------------------------------------------------------------------
-void IRQH_IRQCallBack(BYTE irq)
+void IRQH_IRQCallBack(uint8_t irq)
 {
 	int i;
 
@@ -82,7 +82,7 @@ void IRQH_IRQCallBack(BYTE irq)
 // -----------------------------------------------------------------------
 //   割り込み発生
 // -----------------------------------------------------------------------
-void IRQH_Int(BYTE irq, void* handler)
+void IRQH_Int(uint8_t irq, void* handler)
 {
 	int i;
